поменял почти везде int на long long, может это поможет от оверфлоу

+++ sobolevsky.misha/S5/commands.hpp
@@ -7,9 +7,9 @@
 
 namespace sobolevsky
 {
-  void ascending(std::ostream &out, const sobolevsky::AVLtree< int, std::string, int > &tree);
-  void descending(std::ostream &out, const sobolevsky::AVLtree< int, std::string, int > &tree);
-  void breadth(std::ostream &out, const sobolevsky::AVLtree< int, std::string, int > &tree);
+  void ascending(std::ostream &out, const sobolevsky::AVLtree< long long, std::string, long long > &tree);
+  void descending(std::ostream &out, const sobolevsky::AVLtree< long long, std::string, long long > &tree);
+  void breadth(std::ostream &out, const sobolevsky::AVLtree< long long, std::string, long long > &tree);
 }
 
 #endif
